+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+-------------------------
+ * Add support for anonymous embedded structs, whose fields are loaded as if declared on the config struct
+ * Panic on config structs which can't be loaded, as that's a development error rather than a configuration one
+
 v0.7.0 (2026-08-31)
 -------------------------
  * Fix silently dropped config values when the destination isn't a settable pointer to a struct